\name{rawdata}
\alias{rawdata}
\alias{dbldyeswap}

\docType{data}

\title{Data for a double dye-swap experiment}

\description{
  There are four cDNA arrays comparing two samples in a double dye-swap
  design. 
}

\usage{data(rawdata)}

\format{
  An object of class \code{rawdata}.
}

\examples{
data(rawdata)
}

\keyword{datasets}
