\name{kidney.raw}
\alias{kidney.raw}
\alias{kidney}

\docType{data}

\title{Kidney Data from CAMDA}

\description{
This is a 24-array double reference
design. Six samples are compared to a reference with dye swapped
and all arrays are duplicated. Flag for bad spots is included in the
data.
}

\usage{data(kidney)}
\format{
    An object of class \code{madata}.
}

\source{
  \url{http://www.camda.duke.edu}
}
\references{
  Prichard CC, Hsu L, Delrow J and Nelson PS (2001),
  Project normal: defining normal variance in mouse gene expression,
  \emph{PNAS}, 98:13266
}

\examples{
data(kidney)
}

\keyword{datasets}
