\name{paigen.raw}
\alias{paigen.raw}
\alias{paigen}

\docType{data}

\title{Data for a multiple factor 28-array expeirment}

\description{
This is a multiple factor 28-array experiment.
The experiment is done in Beverly Paigen's Lab in The Jackson Lab.
They took three strains of mice and feed them with two kind of
diets. In that way you get six kind of mice. They picked two
individuals in each group then you have totally 12 distinct mice.
So in this experiment, you have strain, diet and biological replicates
as the factors. You can test the effects from any factor or
any combination of them.
}

\usage{data(paigen)}

\format{
    An object of class \code{rawdata}.
}

\examples{
data(paigen)
}

\keyword{datasets}
