\author{Hao Wu}

\name{Rmaanova.version}
\alias{Rmaanova.version}

\title{Display the current version of the package}

\usage{
Rmaanova.version()
}

\description{
  This is the function to display the current verion number of R/maanova
  pakcage.}

\examples{
Rmaanova.version()
}

\keyword{utilities}
