\name{abf1}
\alias{abf1}
\docType{data}
\title{Data for a 18-array Affymetrix experiment}
\description{
  This is the data set for a 18-array affymetrix experiment. There are
  three mouse strains, AJ, B6 and their F1 offspring. Three biological
  replicates each and two technical replicates for each individual.
}
\usage{data(abf1)}
\format{
  An object of class \code{madata}.
}

\examples{
data(abf1)
}
\keyword{datasets}
